refactor: move onboarding persistence to shared state
