+++ packages/desktop/src/main/index.ts
@@ -14,10 +14,7 @@ import {
 
 import { platform } from '@platform/platform';
 import { SHUTDOWN_PHASE, type LifecycleHost } from '@platform/interfaces';
-import {
-  backfillFirstRunDone,
-  hasUsableSetupCredential,
-} from '@controllers/onboarding/onboardingFunnel';
+import { hasUsableSetupCredential } from '@controllers/onboarding/onboardingFunnel';
 import { getAgent } from '@agent/index/agentRegistry';
 import { registerAgentShutdownHandlers } from '@agent/runtime/agentShutdown';
 import { getServerSideKeyService } from '@auth/serverKeys';
@@ -26,6 +23,7 @@ import type { TerminalRunResult } from '@hosts/uiHosts';
 import { MAIN_VIEW_COMMANDS } from '@shared/ipc';
 import { AgentCategory, agentKeyOf } from '@shared/schemas/agent';
 import { GlobalStateKey } from '@shared/state/stateKeys';
+import { backfillFirstRunDone } from '@shared/state/onboardingState';
 import { setOpenBuildDisplay } from '@tools/approval/latexPreview';
 import { setDesktopAgentResumeHandler } from './desktopAgentResume.js';
 import {
