+++ packages/extension/src/extension.ts
@@ -12,7 +12,6 @@ import { RUNS_STORAGE_DIR } from '@platform/defaults/workspaceStorage';
 import { SHUTDOWN_PHASE, type LifecycleHost } from '@platform/interfaces';
 import { NO_TOOL_AVAILABILITY_HOST } from '@platform/interfaces';
 import { UsageLogService } from '@telemetry/UsageLogService';
-import { backfillFirstRunDone } from '@controllers/onboarding/onboardingFunnel';
 import { seedRosterFromDefaultTeam } from '@controllers/onboarding/defaultTeamSeeding';
 import { defaultSkillSources, setRuntimeSkillSources } from '@skills/index';
 import { StreamLogStore } from '@transcript';
@@ -94,6 +93,7 @@ import { VscodeSecrets } from '@frontend/vscode/vscodeSecrets';
 import { VscodeConfigProvider } from '@frontend/vscode/vscodeConfig';
 import * as logger from '@logger/logUtils';
 import { refreshModelListStateIfNeeded } from '@model/modelListRefresh';
+import { backfillFirstRunDone } from '@shared/state/onboardingState';
 import { migrateLegacyGlobalBashApprovalOverride } from '@shared/settingsView/handlers/approvalHandlers';
 import { GlobalStateKey } from '@shared/state/stateKeys';
 import { setOpenPdfOpener } from '@tools/OpenPdfTool';
